renamed StocksPaginated to stocksPage and moved this into pages folder.
