+++ src/App.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { ThemeProvider } from "@/components/theme-provider";
 import Homepage from "@/pages/homepage";
 import ModeToggle from "@/components/ui/mode-toggle";
-import StocksPaginated from "@/components/divex/StocksPaginated"
+import StocksPage from "@/pages/stocksPage"
 import Login from "@/pages/login";
 import Register from "@/pages/register";
 import CalendarPage from "@/pages/calendarPage";
@@ -32,7 +32,7 @@ function App() {
             <Route path="/account" element={<p>account</p>} />
             <Route path="/portfolio" element={<p>portfolio</p>} />
             <Route path="/calendar" element={<CalendarPage />} />
-            <Route path="/stocks" element={<StocksPaginated />} />
+            <Route path="/stocks" element={<StocksPage />} />
 
           </Routes>
